+++ js/main.js
@@ -71,9 +71,7 @@ function onSubmitHandler(e) {
 	active = 0;
 	updateProgress();
 	registerModal.classList.toggle('modal-overlay__hidden');
-	form.reset();
-	console.log(active, steps.length);
-	
+	form.reset();	
 }
 
 
@@ -82,19 +80,22 @@ function saveUser() {
 	const email = form.querySelector('.email').value;
 	const password = form.querySelector('.password-input').value;
 
-	 let userObj = {
-		 username,
-		 email,
-		 password,
+	let usersArr = JSON.parse(localStorage.getItem('users')) || [];
+
+	let exist =
+		usersArr.length &&
+		JSON.parse(localStorage.getItem('users')).some(
+			(data) => data.email == email
+		);
+	if (!exist) {
+		usersArr.push({ username: username, email: email, password: password });
+		localStorage.setItem('users', JSON.stringify(usersArr));
+	} else {
+		alert('You are registered');
 	}
 
-	let usersArr = JSON.parse(localStorage.getItem('users')) || [];	
-	usersArr.push(userObj);
-	console.log(usersArr, userObj);
-	
-	localStorage.setItem('users', JSON.stringify(usersArr));
-	
- }
+	console.log(usersArr);
+}
 
 
 nextBtn.addEventListener('click', validatePage);
